Move files from spatial to nestkernel to make headers discoverable by extension modules
