+++ nestkernel/grid_mask.h
