+++ nestkernel/layer.cpp
