+++ nestkernel/layer.h
