+++ nestkernel/layer_impl.h
