+++ nestkernel/mask.cpp
