+++ nestkernel/mask_impl.h
