+++ nestkernel/ntree_impl.h
